+++ src/sphWarpCore/operations/bak/wp_gradient_scalar.py
@@ -20,14 +20,14 @@ def computeSPHGradientScalar_Func(
     mode_uint: wp.uint32, kernel_int: wp.int32, gradientMode_int: wp.int32,
     
     neighborList: wp.array(dtype = wp.int64),
-    neighborOffset : wp.int64, numNeighs: wp.int32,
+    neighborOffset : wp.int32, numNeighs: wp.int32,
     
     outputValue: vector(length=Any, dtype=wp.float32)
 ):
     grad_f_interpolated = type(outputValue)(0.0)
     
     for neighborIndex in range(numNeighs):
-        j = wp.int32(neighborList[neighborOffset + wp.int64(neighborIndex)])
+        j = wp.int32(neighborList[neighborOffset + neighborIndex])
         
         mj = masses[j]
         rhoj = densities[j]
@@ -58,7 +58,7 @@ def computeSPHGradientScalar_Kernel(
     domainMin : wp.array(dtype = wp.float32), domainMax : wp.array(dtype = wp.float32), periodicity : wp.array(dtype = wp.bool),
     
     mode_uint: wp.uint32, kernel_int : wp.int32, gradientMode_int: wp.int32,
-    neighborList: wp.array(dtype = wp.int64), neighborListRowOffsets: wp.array(dtype = wp.int64), numNeighbors: wp.array(dtype = wp.int64),
+    neighborList: wp.array(dtype = wp.int64), neighborListRowOffsets: wp.array(dtype = wp.int32), numNeighbors: wp.array(dtype = wp.int32),
     
     outputValues : wp.array(dtype = vector(length=Any, dtype = wp.float32))
 ):                                                                                    
@@ -78,7 +78,7 @@ def computeSPHGradientScalar_Kernel(
         
         periodicity, domainMin, domainMax, 
         mode_uint, kernel_int, gradientMode_int,
-        neighborList, neighborListRowOffsets[i], wp.int32(numNeighbors[i]), 
+        neighborList, neighborListRowOffsets[i], numNeighbors[i], 
         type(outputValues[i])(0.0))
